Turning change_tag_def store to a service
This will be useful when reading the table.

Bug: T194162
