+++ tests/phpunit/includes/changetags/ChangeTagsTest.php
@@ -1,5 +1,7 @@
 <?php
 
+use MediaWiki\MediaWikiServices;
+
 /**
  * @covers ChangeTags
  * @group Database
@@ -505,6 +507,7 @@ class ChangeTagsTest extends MediaWikiTestCase {
 		$dbw = wfGetDB( DB_MASTER );
 		$dbw->delete( 'change_tag', '*' );
 		$dbw->delete( 'change_tag_def', '*' );
+		MediaWikiServices::getInstance()->resetServiceForTesting( 'ChangeTagDefStore' );
 
 		$rcId = 123;
 		ChangeTags::updateTags( [ 'tag1', 'tag2' ], [], $rcId );
